JAMES-1854 Move mailets related with local delivery in same package
