+++ server/mailet/mailets/src/main/java/org/apache/james/transport/mailets/delivery/ToRecipientFolder.java
@@ -16,14 +16,12 @@
  * specific language governing permissions and limitations      *
  * under the License.                                           *
  ****************************************************************/
-
-package org.apache.james.transport.mailets;
+package org.apache.james.transport.mailets.delivery;
 
 import com.google.common.collect.Iterators;
-import com.google.common.collect.Lists;
-import org.apache.james.domainlist.api.DomainList;
 import org.apache.james.mailbox.MailboxManager;
 import org.apache.james.sieverepository.api.SieveRepository;
+import org.apache.james.transport.mailets.ResourceLocatorImpl;
 import org.apache.james.user.api.UsersRepository;
 import org.apache.mailet.Mail;
 import org.apache.mailet.MailetConfig;
@@ -33,23 +31,38 @@ import org.apache.mailet.base.GenericMailet;
 import javax.inject.Inject;
 import javax.inject.Named;
 import javax.mail.MessagingException;
+import java.util.Arrays;
 import java.util.Iterator;
 
 /**
- * Receives a Mail from the Queue and takes care of delivery of the
- * message to local inboxes.
+ * Receives a Mail from the Queue and takes care to deliver the message
+ * to a defined folder of the recipient(s).
  * 
- * This mailet is a composition of RecipientRewriteTable, SieveMailet 
- * and MailboxManager configured to mimic the old "LocalDelivery"
- * James 2.3 behavior.
+ * You have to define the folder name of the recipient(s).
+ * The flag 'consume' will tell is the mail will be further
+ * processed by the upcoming processor mailets, or not.
+ * 
+ * <pre>
+ * &lt;mailet match="RecipientIsLocal" class="ToRecipientFolder"&gt;
+ *    &lt;folder&gt; <i>Junk</i> &lt;/folder&gt;
+ *    &lt;consume&gt; <i>false</i> &lt;/consume&gt;
+ * &lt;/mailet&gt;
+ * </pre>
+ * 
  */
-public class LocalDelivery extends GenericMailet {
-    
-    private org.apache.james.rrt.api.RecipientRewriteTable rrt;
-    private UsersRepository usersRepository;
+public class ToRecipientFolder extends GenericMailet {
+
+    public static final String FOLDER_PARAMETER = "folder";
+    public static final String CONSUME_PARAMETER = "consume";
+
     private MailboxManager mailboxManager;
-    private DomainList domainList;
     private SieveRepository sieveRepository;
+    private UsersRepository usersRepository;
+
+    @Inject
+    public void setMailboxManager(@Named("mailboxmanager")MailboxManager mailboxManager) {
+        this.mailboxManager = mailboxManager;
+    }
 
     @Inject
     public void setSieveRepository(SieveRepository sieveRepository) {
@@ -57,67 +70,31 @@ public class LocalDelivery extends GenericMailet {
     }
 
     @Inject
-    public void setRrt(org.apache.james.rrt.api.RecipientRewriteTable rrt) {
-        this.rrt = rrt;
-    }
-
-    @Inject
     public void setUsersRepository(UsersRepository usersRepository) {
         this.usersRepository = usersRepository;
     }
-    
-    @Inject
-    public void setMailboxManager(@Named("mailboxmanager") MailboxManager mailboxManager) {
-        this.mailboxManager = mailboxManager;
-    }
-    
-    @Inject
-    public void setDomainList(DomainList domainList) {
-        this.domainList = domainList;
-    }
 
     private SieveMailet sieveMailet;  // Mailet that actually stores the message
-    private RecipientRewriteTable recipientRewriteTable;  // Mailet that applies RecipientRewriteTable
 
     /**
-     * Delivers a mail to a local mailbox.
+     * Delivers a mail to a local mailbox in a given folder.
      * 
-     * @param mail the mail being processed
-     * 
-     * @throws MessagingException if an error occurs while storing the mail
+     * @see org.apache.mailet.base.GenericMailet#service(org.apache.mailet.Mail)
      */
+    @Override
     public void service(Mail mail) throws MessagingException {
-        recipientRewriteTable.service(mail);
         if (!mail.getState().equals(Mail.GHOST)) {
             sieveMailet.service(mail);
         }
     }
 
-    /**
-     * Return a string describing this mailet.
-     * 
-     * @return a string describing this mailet
-     */
-    public String getMailetInfo() {
-        return "Local Delivery Mailet";
-    }
-
-    /**
-     * @see org.apache.mailet.base.GenericMailet#init()
-     */
+    @Override
     public void init() throws MessagingException {
-        
         super.init();
-
-        recipientRewriteTable = new RecipientRewriteTable();
-        recipientRewriteTable.setDomainList(domainList);
-        recipientRewriteTable.setRecipientRewriteTable(rrt);
-        recipientRewriteTable.init(getMailetConfig());
-        sieveMailet = new SieveMailet(usersRepository,
-            mailboxManager,
-            ResourceLocatorImpl.instanciate(usersRepository, sieveRepository),
-            "INBOX");
+        sieveMailet = new SieveMailet(usersRepository, mailboxManager, ResourceLocatorImpl.instanciate(usersRepository, sieveRepository), getInitParameter(FOLDER_PARAMETER, "INBOX"));
         sieveMailet.init(new MailetConfig() {
+            
+            @Override
             public String getInitParameter(String name) {
                 if ("addDeliveryHeader".equals(name)) {
                     return "Delivered-To";
@@ -128,23 +105,30 @@ public class LocalDelivery extends GenericMailet {
                 }
             }
 
+            @Override
             public Iterator<String> getInitParameterNames() {
-                return Iterators.concat(
-                        getMailetConfig().getInitParameterNames(), 
-                        Lists.newArrayList("addDeliveryHeader", "resetReturnPath").iterator());
+                return Iterators.concat(getMailetConfig().getInitParameterNames(),
+                        Arrays.asList("addDeliveryHeader", "resetReturnPath").iterator());
             }
             
+            @Override
             public MailetContext getMailetContext() {
                 return getMailetConfig().getMailetContext();
             }
 
+            @Override
             public String getMailetName() {
                 return getMailetConfig().getMailetName();
             }
 
         });
-        // Override the default value of "quiet"
         sieveMailet.setQuiet(getInitParameter("quiet", true));
+        sieveMailet.setConsume(getInitParameter(CONSUME_PARAMETER, false));
+    }
+
+    @Override
+    public String getMailetInfo() {
+        return ToRecipientFolder.class.getName() + " Mailet";
     }
 
 }
